refactor: extract profile events from AgentWorkspace
